Fix type annotations for Sets Java port
This module had some incorrect type annotations. In general, the class declaration should look like:

```
public class MyFunc<A> extends PrimitiveFunction<A>
```

And typeclass A should not be used anywhere else within the class. This is because A is only meant to act as an annotation to help convert the function into a Term.